+++ lib/clean_framework.dart
@@ -8,5 +8,7 @@ export 'package:clean_framework/src/feature_state/feature_mapper.dart';
 export 'package:clean_framework/src/feature_state/feature_state_provider.dart';
 export 'package:clean_framework/src/feature_state/feature_widget.dart';
 export 'package:clean_framework/src/logger.dart';
+export 'package:clean_framework/src/open_feature/open_feature.dart';
 export 'package:clean_framework/src/routing/app_router.dart';
+export 'package:clean_framework/src/widgets/widgets.dart';
 export 'package:either_dart/either.dart';