+++ lib/src/open_feature/src/core/enums/error_code.dart
@@ -1,0 +1,13 @@
+// coverage:ignore-file
+
+enum ErrorCode {
+  providerNotReady('PROVIDER_NOT_READY'),
+  flagNotFound('FLAG_NOT_FOUND'),
+  parseError('PARSE_ERROR'),
+  typeMismatch('TYPE_MISMATCH'),
+  general('GENERAL');
+
+  const ErrorCode(this.rawCode);
+
+  final String rawCode;
+}