+++ lib/src/open_feature/src/core/enums/flag_value_type.dart
@@ -1,0 +1,6 @@
+enum FlagValueType {
+  boolean,
+  string,
+  number,
+  object,
+}