+++ lib/src/open_feature/src/evaluation_context/context_transformation_mixin.dart
@@ -1,0 +1,8 @@
+import 'dart:async';
+
+import '../provider/feature_provider.dart';
+import 'evaluation_context.dart';
+
+mixin ContextTransformationMixin on FeatureProvider {
+  FutureOr<EvaluationContext> transformContext(EvaluationContext context);
+}