+++ lib/src/open_feature/src/exceptions/src/type_mismatch_exception.dart
@@ -1,0 +1,7 @@
+// coverage:ignore-file
+
+import '../open_feature_exception.dart';
+
+class TypeMismatchException extends OpenFeatureException {
+  TypeMismatchException(super.message) : super(code: ErrorCode.typeMismatch);
+}