+++ lib/src/open_feature/src/flag_evaluation/flag_evaluation_options.dart
@@ -1,0 +1,11 @@
+import '../hook/hook.dart';
+
+class FlagEvaluationOptions {
+  const FlagEvaluationOptions({
+    this.hooks = const [],
+    this.hookHints,
+  });
+
+  final List<Hook> hooks;
+  final HookHints? hookHints;
+}