+++ lib/src/tests/test_helpers.dart
@@ -65,6 +65,7 @@ void uiTest(
   dynamic tags,
   Size? screenSize,
   Iterable<LocalizationsDelegate>? localizationDelegates,
+  Widget Function(Widget)? parentBuilder,
 }) {
   assert(
     () {
@@ -130,7 +131,10 @@ void uiTest(
         child = _scopedChild(builder!());
       }
 
-      await tester.pumpWidget(child, pumpDuration);
+      await tester.pumpWidget(
+        parentBuilder == null ? child : parentBuilder(child),
+        pumpDuration,
+      );
 
       if (postFrame == null) {
         await tester.pumpAndSettle();